+++ src/utils/Auth.js
@@ -1,5 +1,6 @@
 import sha256 from 'crypto-js/sha256';
 import ConfigAccumulator from '@/utils/ConfigAccumalator';
+import ErrorHandler from '@/utils/ErrorHandler';
 import { cookieKeys, localStorageKeys, userStateEnum } from '@/utils/defaults';
 
 /* Uses config accumulator to get and return app config */
@@ -9,10 +10,48 @@ const getAppConfig = () => {
   return config.appConfig || {};
 };
 
-/* Returns the users array from appConfig, if available, else an empty array */
+/**
+ * Called when the user is still using array for users, prints warning
+ * This was a breaking change, implemented in V 1.6.5
+ * Support for old user structure will be removed in V 1.7.0
+ */
+const printWarning = () => {
+  const msg = 'From V 1.6.5 onwards, the structure of the users object has changed.';
+  // eslint-disable-next-line no-console
+  console.warn(msg);
+};
+
+/* Returns true if keycloak is enabled */
+export const isKeycloakEnabled = () => {
+  const appConfig = getAppConfig();
+  if (!appConfig.auth) return false;
+  return appConfig.auth.enableKeycloak || false;
+};
+
+/* Returns the users keycloak config */
+export const getKeycloakConfig = () => {
+  const appConfig = getAppConfig();
+  if (!isKeycloakEnabled()) return false;
+  const { keycloak } = appConfig.auth;
+  const { serverUrl, realm, clientId } = keycloak;
+  if (!serverUrl || !realm || !clientId) {
+    ErrorHandler('Keycloak config missing- please ensure you specify: serverUrl, realm, clientId');
+    return false;
+  }
+  return keycloak;
+};
+
+/* Returns array of users from appConfig.auth, if available, else an empty array */
 const getUsers = () => {
   const appConfig = getAppConfig();
-  return appConfig.auth || [];
+  const auth = appConfig.auth || {};
+  // Check if the user is still using previous schema type
+  if (Array.isArray(auth)) {
+    printWarning(); // Print warning message
+    return auth; // Let the user proceed anyway, will remove in V 1.7.0
+  }
+  // Otherwise, return the users array, if available
+  return auth.users || [];
 };
 
 /**
@@ -58,7 +97,15 @@ export const isAuthEnabled = () => {
 /* Returns true if guest access is enabled */
 export const isGuestAccessEnabled = () => {
   const appConfig = getAppConfig();
-  return appConfig.enableGuestAccess || false;
+  if (appConfig.enableGuestAccess) {
+    // User is still using the old auth method
+    printWarning();
+    return true;
+  }
+  if (appConfig.auth && !Array.isArray(appConfig.auth)) {
+    return appConfig.auth.enableGuestAccess || false;
+  }
+  return false;
 };
 
 /**